chore: move parsing to separate mod
