+++ src/fetch.rs
@@ -1,6 +1,6 @@
 use crate::{
+    parsing::{parse_cargo_test, Text},
     prettify::make_pretty,
-    regex::{TestInfo, Text},
 };
 use std::process::{Command, Output};
 use termtree::Tree;
@@ -31,7 +31,7 @@ pub fn cargo_test() -> Output {
 
 pub fn parse_cargo_test_output<'s>(stderr: &'s str, stdout: &'s str) -> Tree<Text<'s>> {
     let mut tree = Tree::new("Generated by cargo-pretty-test");
-    for (pkg, data) in TestInfo::parse_cargo_test(stderr, stdout).pkgs {
+    for (pkg, data) in parse_cargo_test(stderr, stdout).pkgs {
         tree.push(Tree::new(pkg.unwrap_or("tests")).with_leaves(
             data.inner.into_iter().filter_map(|data| {
                 let parsed = data.info.parsed;
