+++ tests/parsing.rs
@@ -1,4 +1,4 @@
-use cargo_pretty_test::regex::{TestInfo, TestType};
+use cargo_pretty_test::parsing::{parse_cargo_test_with_empty_ones, TestType};
 use pretty_assertions::assert_eq;
 use std::time::Duration;
 
@@ -44,7 +44,7 @@ test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; fini
 fn parse_stderr_stdout() {
     use TestType::*;
 
-    let parsed_tests_info = TestInfo::parse_cargo_test_with_empty_ones(STDERR, STDOUT);
+    let parsed_tests_info: Vec<_> = parse_cargo_test_with_empty_ones(STDERR, STDOUT).collect();
     let parsed_stderr = parsed_tests_info
         .iter()
         .map(|(r, _)| (r.ty, r.src.src_path, r.src.bin_name))
@@ -64,7 +64,7 @@ fn parse_stderr_stdout() {
     assert!(
         (parsed_tests_info
             .iter()
-            .map(|(_, v)| v.stat.finished_in)
+            .map(|(_, v)| v.stats.finished_in)
             .sum::<Duration>()
             .as_secs_f32()
             - total_time)
@@ -83,7 +83,7 @@ fn parse_stderr_stdout() {
                 r.src.src_path,
                 r.src.bin_name,
                 &i.parsed.tree,
-                &i.stat
+                &i.stats
             ))
             .collect::<Vec<_>>(),
     );
